Move FromHashable
