+++ ruby/optify/rbi/optify.rbi
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 # typed: strict
 
 # Tools for working with configurations declared in files.
@@ -10,15 +11,15 @@ module Optify
     # @param feature_names [Array<String>] The enabled feature names to use to build the options.
     # @param config_class [ConfigType] The class of the configuration to return.
     # @return [ConfigType] The options.
-    sig {
+    sig do
       type_parameters(:Config)
-      .params(
-        key: String,
-        feature_names: T::Array[String],
-        config_class: T::Class[T.type_parameter(:Config)]
-      )
-      .returns(T.type_parameter(:Config))
-    }
+        .params(
+          key: String,
+          feature_names: T::Array[String],
+          config_class: T::Class[T.type_parameter(:Config)]
+        )
+        .returns(T.type_parameter(:Config))
+    end
     def get_options(key, feature_names, config_class); end
 
     # Fetches options in JSON format based on the provided key and feature names.
