+++ ruby/optify/test/my_config.rb
@@ -4,40 +4,7 @@
 require 'sorbet-runtime'
 require 'tapioca'
 
-class FromHashable
-  extend T::Sig
-
-  # Create a new instance of the class from a hash.
-  #
-  # @param hash [Hash] The hash to create the instance from.
-  # @return The new instance.
-  sig { params(hash: T::Hash[T.untyped, T.untyped]).returns(T.attached_class) }
-  def self.from_hash(hash)
-    result = new
-
-    hash.each do |key, value|
-      # TODO: Might need some error handling here, but it should be fine if type signatures are used.
-      # TODO Handle nillable types.
-      case value
-      when Array
-        sig_return_type = T::Utils.signature_for_method(instance_method(key)).return_type
-        inner_type = sig_return_type.type.raw_type
-        value = value.map { |v| inner_type.from_hash(v) } if inner_type.respond_to?(:from_hash)
-      when Hash
-        sig_return_type = T::Utils.signature_for_method(instance_method(key)).return_type
-        if sig_return_type.respond_to?(:raw_type)
-          type_for_key = sig_return_type.raw_type
-          value = type_for_key.from_hash(value) if type_for_key.respond_to?(:from_hash)
-        end
-      end
-
-      result.instance_variable_set("@#{key}", value)
-    end
-    result
-  end
-end
-
-class MyObject < FromHashable
+class MyObject < Optify::FromHashable
   private
 
   attr_writer :one, :two, :string, :deeper
@@ -58,7 +25,7 @@ class MyObject < FromHashable
 end
 
 # A custom configuration for testing.
-class MyConfig < FromHashable
+class MyConfig < Optify::FromHashable
   private
 
   attr_writer :rootString, :myArray, :myObject, :myObjects, :deeper # rubocop:disable Naming/MethodName
